refactor(server): extract route families
